refactor: MenuGroup 관련 파일들 Menu 패키지로 이동
