--- src/main/java/kitchenpos/menuGroup/application/MenuGroupService.java
@@ -1,33 +1,0 @@
-package kitchenpos.menuGroup.application;
-
-import kitchenpos.menuGroup.domain.MenuGroup;
-import kitchenpos.menuGroup.domain.MenuGroupRepository;
-import kitchenpos.menuGroup.dto.MenuGroupRequest;
-import kitchenpos.menuGroup.dto.MenuGroupResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-public class MenuGroupService {
-    private final MenuGroupRepository menuGroupRepository;
-
-    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
-        this.menuGroupRepository = menuGroupRepository;
-    }
-
-    @Transactional
-    public MenuGroupResponse create(final MenuGroupRequest menuGroupRequest) {
-        MenuGroup menuGroup = MenuGroup.of(menuGroupRequest.getName());
-        menuGroup = menuGroupRepository.save(menuGroup);
-        return MenuGroupResponse.of(menuGroup);
-    }
-
-    public List<MenuGroupResponse> list() {
-        return menuGroupRepository.findAll().stream()
-                .map(MenuGroupResponse::of)
-                .collect(Collectors.toList());
-    }
-}
